+++ src/ProSuite.Microservices.Client.AGP/GeometryProcessing/RemoveOverlaps/RemoveOverlapsClientUtils.cs
@@ -74,7 +74,7 @@ namespace ProSuite.Microservices.Client.AGP.GeometryProcessing.RemoveOverlaps
 			CalculateOverlapsRequest request =
 				CreateCalculateOverlapsRequest(selectedFeatures, overlappingFeatures, inExtent);
 
-			int deadline = FeatureProcessingUtils.GetProcessingTimeout(selectedFeatures.Count);
+			long deadline = FeatureProcessingUtils.GetProcessingTimeout(selectedFeatures.Count);
 
 			CalculateOverlapsResponse response =
 				GrpcClientUtils.Try(
@@ -144,7 +144,7 @@ namespace ProSuite.Microservices.Client.AGP.GeometryProcessing.RemoveOverlaps
 				selectedFeatures, overlapsToRemove, overlappingFeatures, options,
 				out updateFeatures);
 
-			int deadline =
+			long deadline =
 				FeatureProcessingUtils.GetProcessingTimeout(request.SourceFeatures.Count);
 
 			RemoveOverlapsResponse response =
